LayoutTests/fast/profiler/multiple-and-different-scoped-function-calls.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
if (window.internals)
    internals.setLegacyJavaScriptProfilingEnabled(true);

function startTest()
{
    console.profile("Multiple and different scoped calls to the same function");

    insertNewText();
    insertNewText();

    intermediaryFunction();

    endTest();
}
</script>
</head>

<body onload="startTest()">
This page's JavaScript calls the same function from multiple and differently scoped locations.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be two calls to insertNewText under
startTest() and one under an (anonymous function) under intermediaryFunction().
<div id="output"></div>
</body>
</html>
