LayoutTests/fast/profiler/multiple-and-different-scoped-function-calls-expected.txt

This page's JavaScript calls the same function from multiple and differently scoped locations. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be two calls to insertNewText under startTest() and one under an (anonymous function) under intermediaryFunction().

Profile title: Multiple and different scoped calls to the same function
Thread_1 (no file) (line 0)
   startTest multiple-and-different-scoped-function-calls.html (line 11)
      insertNewText profiler-test-JS-resources.js (line 17)
         createElement (no file) (line 0)
         createTextNode (no file) (line 0)
         appendChild (no file) (line 0)
         getElementById (no file) (line 0)
      intermediaryFunction profiler-test-JS-resources.js (line 33)
         anonymousFunction profiler-test-JS-resources.js (line 29)
            insertNewText profiler-test-JS-resources.js (line 17)
               createElement (no file) (line 0)
               createTextNode (no file) (line 0)
               appendChild (no file) (line 0)
               getElementById (no file) (line 0)
      endTest profiler-test-JS-resources.js (line 1)
